feat: Add fire suppression master arm control
Introduces a master arm switch to the fire suppression system. This switch must be engaged along with the fire handle being pulled to allow bottle discharge. This adds an extra layer of safety to prevent accidental discharge.

The `masterArmed` state has been added to `FireSystemState` and is now managed in `useEngineSimulation`. A corresponding `toggleFireMaster` function and `onToggleMaster` prop have been added to `FirePanel` and `App.tsx` respectively.